Seperated test classes into smaller parts modified hooks, and page steps, implemented groups to each of them implemented IRetryAnalyzer interface modified get(browser) environment made it nullable to return chrome as default
