+++ Extent_Insider - Page Functionality.html
@@ -35,7 +35,7 @@
 <a href="#"><span class="badge badge-primary"></span></a>
 </li>
 <li class="m-r-10">
-<a href="#"><span class="badge badge-primary">May 6, 2023 02:18:02 AM</span></a>
+<a href="#"><span class="badge badge-primary">May 7, 2023 03:31:53 AM</span></a>
 </li>
 </ul>
 </div>
@@ -81,7 +81,7 @@
           <div class="test-detail">
             <p class="name">homePageTest</p>
             <p class="text-sm">
-              <span>02:18:03 AM</span> / <span>00:00:14:105</span>
+              <span>03:31:53 AM</span> / <span>00:00:06:522</span>
               <span class="badge pass-bg log float-right">Pass</span>
             </p>
           </div>
@@ -90,9 +90,9 @@
 <div class="p-v-10">
 <div class="info">
 <h5 class="test-status text-pass">homePageTest</h5>
-<span class='badge badge-success'>05.06.2023 02:18:03</span>
-<span class='badge badge-danger'>05.06.2023 02:18:17</span>
-<span class='badge badge-default'>00:00:14:105</span>
+<span class='badge badge-success'>05.07.2023 03:31:53</span>
+<span class='badge badge-danger'>05.07.2023 03:32:00</span>
+<span class='badge badge-default'>00:00:06:522</span>
 &middot; <span class='uri-anchor badge badge-default'>#test-id=1</span>
 </div>
 </div>
@@ -101,15 +101,8 @@
   <thead><tr><th class="status-col">Status</th><th class="timestamp-col">Timestamp</th><th class="details-col">Details</th></tr></thead>
   <tbody>
       <tr class="event-row">
-        <td><span class="badge log info-bg">Info</span></td>
-        <td>2:18:03 AM</td>
-        <td>
-          Home page test2 has been started
-        </td>
-      </tr>
-      <tr class="event-row">
         <td><span class="badge log pass-bg">Pass</span></td>
-        <td>2:18:17 AM</td>
+        <td>3:32:00 AM</td>
         <td>
           <span class='badge white-text green'>homePageTest Test Case PASSED</span>
         </td>
@@ -126,7 +119,7 @@
           <div class="test-detail">
             <p class="name">careersPageTest</p>
             <p class="text-sm">
-              <span>02:18:17 AM</span> / <span>00:00:00:920</span>
+              <span>03:32:00 AM</span> / <span>00:00:00:602</span>
               <span class="badge pass-bg log float-right">Pass</span>
             </p>
           </div>
@@ -135,9 +128,9 @@
 <div class="p-v-10">
 <div class="info">
 <h5 class="test-status text-pass">careersPageTest</h5>
-<span class='badge badge-success'>05.06.2023 02:18:17</span>
-<span class='badge badge-danger'>05.06.2023 02:18:18</span>
-<span class='badge badge-default'>00:00:00:920</span>
+<span class='badge badge-success'>05.07.2023 03:32:00</span>
+<span class='badge badge-danger'>05.07.2023 03:32:00</span>
+<span class='badge badge-default'>00:00:00:602</span>
 &middot; <span class='uri-anchor badge badge-default'>#test-id=2</span>
 </div>
 </div>
@@ -146,17 +139,48 @@
   <thead><tr><th class="status-col">Status</th><th class="timestamp-col">Timestamp</th><th class="details-col">Details</th></tr></thead>
   <tbody>
       <tr class="event-row">
-        <td><span class="badge log info-bg">Info</span></td>
-        <td>2:18:17 AM</td>
-        <td>
-          careers page test2 has been started
-        </td>
-      </tr>
-      <tr class="event-row">
         <td><span class="badge log pass-bg">Pass</span></td>
-        <td>2:18:18 AM</td>
+        <td>3:32:00 AM</td>
         <td>
           <span class='badge white-text green'>careersPageTest Test Case PASSED</span>
+        </td>
+      </tr>
+  </tbody>
+</table>
+</div>
+          </div>
+        </li>
+        <li class="test-item"  status="pass" test-id="3"
+          author=""
+          tag=""
+          device="">
+          <div class="test-detail">
+            <p class="name">jobsPageTest</p>
+            <p class="text-sm">
+              <span>03:32:00 AM</span> / <span>00:00:12:057</span>
+              <span class="badge pass-bg log float-right">Pass</span>
+            </p>
+          </div>
+          <div class="test-contents d-none">
+<div class="detail-head">
+<div class="p-v-10">
+<div class="info">
+<h5 class="test-status text-pass">jobsPageTest</h5>
+<span class='badge badge-success'>05.07.2023 03:32:00</span>
+<span class='badge badge-danger'>05.07.2023 03:32:12</span>
+<span class='badge badge-default'>00:00:12:057</span>
+&middot; <span class='uri-anchor badge badge-default'>#test-id=3</span>
+</div>
+</div>
+</div><div class="detail-body mt-4">
+<table class="table table-sm">
+  <thead><tr><th class="status-col">Status</th><th class="timestamp-col">Timestamp</th><th class="details-col">Details</th></tr></thead>
+  <tbody>
+      <tr class="event-row">
+        <td><span class="badge log pass-bg">Pass</span></td>
+        <td>3:32:12 AM</td>
+        <td>
+          <span class='badge white-text green'>jobsPageTest Test Case PASSED</span>
         </td>
       </tr>
   </tbody>
@@ -178,19 +202,19 @@
 <div class="col-md-3">
 <div class="card"><div class="card-body">
 <p class="m-b-0">Started</p>
-<h3>May 6, 2023 02:18:02 AM</h3>
+<h3>May 7, 2023 03:31:53 AM</h3>
 </div></div>
 </div>
 <div class="col-md-3">
 <div class="card"><div class="card-body">
 <p class="m-b-0">Ended</p>
-<h3>May 6, 2023 02:18:18 AM</h3>
+<h3>May 7, 2023 03:32:12 AM</h3>
 </div></div>
 </div>
 <div class="col-md-3">
 <div class="card"><div class="card-body">
 <p class="m-b-0 text-pass">Tests Passed</p>
-<h3>2</h3>
+<h3>3</h3>
 </div></div>
 </div>
 <div class="col-md-3">
@@ -213,7 +237,7 @@
 </div>
 <div class="card-footer">
 <div><small data-tooltip='100%'>
-<b>2</b> tests passed
+<b>3</b> tests passed
 </small>
 </div>
 <div>
@@ -235,10 +259,10 @@
 </div>
 </div>
 <div class="card-footer">
-<div><small data-tooltip='50%'><b>2</b> events passed</small></div>
+<div><small data-tooltip='100%'><b>3</b> events passed</small></div>
 <div>
 <small data-tooltip='0%'><b>0</b> events failed,
-<b data-tooltip='%'>2</b> others
+<b data-tooltip='%'>0</b> others
 </small>
 </div>
 </div>
@@ -254,7 +278,7 @@
 </div></div>
 <script>
 var timeline = {
-"homePageTest":14.105,"careersPageTest":0.92
+"homePageTest":6.522,"careersPageTest":0.602,"jobsPageTest":12.057
 };
 </script>
 <div class="row">
@@ -263,7 +287,7 @@ var timeline = {
 <script>
 var statusGroup = {
 parentCount: 5,
-passParent: 2,
+passParent: 3,
 failParent: 0,
 warningParent: 0,
 skipParent: 0,
@@ -280,11 +304,11 @@ warningGrandChild: 0,
 skipGrandChild: 0,
 infoGrandChild: 0,
 eventsCount: 5,
-passEvents: 2,
+passEvents: 3,
 failEvents: 0,
 warningEvents: 0,
 skipEvents: 0,
-infoEvents: 2
+infoEvents: 0
 };
 </script>        </div>
       </div>
